requirements_nfl.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
argon2-cffi=20.1.0=py38h27cfd23_1
arrow=0.17.0=pypi_0
astunparse=1.6.3=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attention=4.0=pypi_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
binaryornot=0.4.4=pypi_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
bokeh=2.3.1=py38h06a4308_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.1.1=pypi_0
category-encoders=2.2.2=pypi_0
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.5=py38h261ae71_0
chardet=3.0.4=pypi_0
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
cookiecutter=1.7.2=pypi_0
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask=2.30.0=py_0
dask-core=2.30.0=py_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2.30.1=py38h06a4308_0
entrypoints=0.3=py38_0
et-xmlfile=1.1.0=pypi_0
expat=2.3.0=h2531618_2
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.9.0=pyhd3eb1b0_0
gast=0.3.3=pypi_0
glib=2.68.1=h36276a3_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
heapdict=1.0.1=py_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
imbalanced-learn=0.8.0=pypi_0
importlib-metadata=3.10.0=py38h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pypi_0
joblib=0.17.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter-contrib-core=0.3.3=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.4.1=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
keras=2.4.3=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0=py38h612dafd_4
locket=0.2.1=py38h06a4308_1
lxml=4.6.2=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
msgpack-python=1.0.2=py38hff7bd54_1
nb_conda_kernels=2.3.0=py38_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
notebook=6.3.0=py38h06a4308_0
numba=0.53.1=py38ha9443f7_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openpyxl=3.0.7=pypi_0
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
patsy=0.5.1=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
plotly=4.14.1=pypi_0
plotly-orca=1.3.1=1
poyo=0.5.0=pypi_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python-dotenv=0.15.0=pypi_0
python-slugify=4.0.1=pypi_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py38h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
rsa=4.6=pypi_0
scikit-image=0.16.2=py38h0573a6f_0
scikit-learn=0.24.1=py38ha9443f7_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=50.3.2=pypi_0
shap=0.37.0=py38h0ef3d22_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
sklearn=0.0=pypi_0
slicer=0.0.7=pyhd8ed1ab_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
split-folders=0.4.3=pypi_0
sqlite=3.35.4=hdfb4753_0
statsmodels=0.12.1=pypi_0
tbb=2020.3=hfd86e86_0
tblib=1.7.0=py_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.4.1=pypi_0
tensorflow-addons=0.12.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py38h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py38h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typeguard=2.12.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.2=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xgboost=1.4.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zict=2.0.0=pyhd3eb1b0_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
